requirements.txt
asgiref==3.5.0
dj-database-url==0.5.0
Django==4.0.2
django-cors-headers==3.11.0
django-environ==0.8.1
django-heroku==0.3.1
dnspython==2.2.0
gunicorn==20.1.0
numpy==1.22.2
pandas==1.4.1
psycopg2==2.9.3
pymongo==4.0.1
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
whitenoise==6.0.0
